Moved recorders and related classes into their own package
